Move delete button to page header
