+++ changelogs/unreleased/749-GuessWhoSamFoo
@@ -1,0 +1,1 @@
+Moved delete button next to page header
