+++ internal/printer/object_test.go
@@ -10,15 +10,11 @@ import (
 	"testing"
 
 	"github.com/golang/mock/gomock"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 
-	"github.com/vmware-tanzu/octant/internal/octant"
-	"github.com/vmware-tanzu/octant/pkg/action"
 	"github.com/vmware-tanzu/octant/pkg/plugin/fake"
-	"github.com/vmware-tanzu/octant/pkg/store"
 
 	"github.com/vmware-tanzu/octant/internal/testutil"
 	"github.com/vmware-tanzu/octant/pkg/plugin"
@@ -145,20 +141,6 @@ func Test_Object_ToComponent(t *testing.T) {
 			},
 		},
 		{
-			name:   "add button",
-			object: deployment,
-			initFunc: func(object *Object, options *initOptions) {
-				stubPlugins(options.PluginPrinter)
-				object.AddButton("button name", action.Payload{})
-			},
-			buttons: []component.Button{
-				component.NewButton("button name", action.Payload{}),
-			},
-			sections: []component.FlexLayoutSection{
-				defaultConfigSection,
-			},
-		},
-		{
 			name:   "register items",
 			object: deployment,
 			initFunc: func(o *Object, options *initOptions) {
@@ -271,44 +253,8 @@ func Test_Object_ToComponent(t *testing.T) {
 
 			expected := component.NewFlexLayout("Summary")
 			expected.AddSections(tc.sections...)
-			buttonGroup := component.NewButtonGroup()
-			if len(tc.buttons) > 0 {
-				for _, button := range tc.buttons {
-					buttonGroup.AddButton(button)
-				}
-			}
-
-			key, err := store.KeyFromObject(deployment)
-			require.NoError(t, err)
-
-			buttonGroup.AddButton(
-				component.NewButton("Delete",
-					action.CreatePayload(octant.ActionDeleteObject, key.ToActionPayload()),
-					component.WithButtonConfirmation(
-						"Delete Deployment",
-						"Are you sure you want to delete *Deployment* **deployment**? This action is permanent and cannot be recovered.",
-					)))
-			expected.SetButtonGroup(buttonGroup)
 
 			component.AssertEqual(t, expected, got)
 		})
 	}
-}
-
-func Test_deleteObjectConfirmation(t *testing.T) {
-	pod := testutil.CreatePod("pod")
-	option, err := deleteObjectConfirmation(pod)
-	require.NoError(t, err)
-
-	button := component.Button{}
-	option(&button)
-
-	expected := component.Button{
-		Confirmation: &component.Confirmation{
-			Title: "Delete Pod",
-			Body:  "Are you sure you want to delete *Pod* **pod**? This action is permanent and cannot be recovered.",
-		},
-	}
-
-	assert.Equal(t, expected, button)
 }
