+++ web/src/app/modules/shared/components/presentation/button-group/button-group.component.scss
@@ -1,0 +1,3 @@
+.btn-primary {
+    margin-right: 0rem;
+}
