+++ web/src/app/modules/shared/components/presentation/tabs/tabs.component.html
@@ -1,9 +1,18 @@
-<h2>
-    <ng-container *ngIf="iconName">
-        <clr-icon [attr.shape]="iconName" size="24"></clr-icon>
-    </ng-container>
-    {{ title }}
-</h2>
+<div class="clr-row clr-align-items-center clr-justify-content-between">
+    <div class="clr-col-10">
+        <h2>
+            <ng-container *ngIf="iconName">
+                <clr-icon [attr.shape]="iconName" size="24"></clr-icon>
+            </ng-container>
+            {{ title }}
+        </h2>
+    </div>
+    <div class="clr-col-2">
+        <div class="action-buttons">
+            <app-button-group [view]="buttonGroup"></app-button-group>
+        </div>
+    </div>
+</div>
 <clr-tabs *ngIf="tabs.length > 0">
     <clr-tab *ngFor="let tab of tabs; trackBy: identifyTab">
         <div clrTabLink class="tab-button">
