+++ web/src/app/modules/shared/components/presentation/tabs/tabs.component.scss
@@ -5,3 +5,7 @@
 .tab-button {
   outline: none;
 }
+
+.action-buttons {
+  float: right;
+}
