+++ web/src/app/modules/shared/components/presentation/tabs/tabs.component.ts
@@ -9,7 +9,7 @@ import {
   SimpleChanges,
 } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { View } from 'src/app/modules/shared/models/content';
+import { View, ButtonGroupView } from 'src/app/modules/shared/models/content';
 import { SliderService } from 'src/app/modules/shared/slider/slider.service';
 import { ViewService } from '../../../services/view/view.service';
 import { WebsocketService } from '../../../services/websocket/websocket.service';
@@ -33,6 +33,7 @@ export class TabsComponent implements OnChanges, OnInit {
   @Input() iconName: string;
   @Input() closable: boolean;
   @Input() extView: boolean;
+  @Input() buttonGroup: ButtonGroupView;
 
   tabs: Tab[] = [];
   activeTab: string;
