+++ web/src/app/modules/sugarloaf/components/smart/content/content.component.html
@@ -2,7 +2,7 @@
     <ng-container *ngIf="hasReceivedContent">
         <ng-container *ngIf="hasTabs; then withTabs; else withoutTabs"></ng-container>
         <ng-template #withTabs>
-            <app-object-tabs [iconName]="iconName" [title]="title" [views]="views"></app-object-tabs>
+            <app-object-tabs [buttonGroup]="buttonGroup" [iconName]="iconName" [title]="title" [views]="views"></app-object-tabs>
         </ng-template>
         <ng-template #withoutTabs>
             <app-content-switcher [view]="singleView"></app-content-switcher>
